requirements.txt
reactpy==1.0.1
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
aiohttp==3.8.4
selenium==4.10.0
undetected-chromedriver==3.5.0
PyAutoGUI==0.9.54
reactpy[starlette]